Make completer complete commands at beginning of S-expressions
